Break the SerilogExpressions dependency, unify Seq.Syntax references to the shipped one
