+++ global.json
@@ -1,5 +1,5 @@
 {
   "sdk": {
-    "version": "8.0.101"
+    "version": "8.0.100"
   }
 }
