+++ src/SeqCli/SeqCli.csproj
@@ -29,7 +29,6 @@
     <PackageReference Include="Destructurama.JsonNet" Version="2.0.1" />
     <PackageReference Include="newtonsoft.json" Version="13.0.3" />
     <PackageReference Include="Serilog" Version="3.1.1" />
-    <PackageReference Include="serilog.expressions" Version="4.0.0" />
     <PackageReference Include="Serilog.Formatting.Compact" Version="2.0.0" />
     <PackageReference Include="Serilog.Formatting.Compact.Reader" Version="3.0.0" />
     <PackageReference Include="Serilog.Sinks.Console" Version="5.0.1" />
