+++ src/SeqCli/Syntax/SeqSyntax.cs
@@ -1,5 +1,4 @@
-﻿using System.Diagnostics.CodeAnalysis;
-using Seq.Syntax.Expressions;
+﻿using Seq.Syntax.Expressions;
 
 namespace SeqCli.Syntax;
 
@@ -8,21 +7,5 @@ static class SeqSyntax
     public static CompiledExpression CompileExpression(string expression)
     {
         return SerilogExpression.Compile(expression, nameResolver: new SeqCliNameResolver());
-    }
-}
-
-class SeqCliNameResolver: NameResolver
-{
-    public override bool TryResolveBuiltInPropertyName(string alias, [MaybeNullWhen(false)] out string target)
-    {
-        switch (alias)
-        {
-            case "@l":
-                target = "coalesce(SeqCliOriginalLevel, @l)";
-                return true;
-            default:
-                target = null;
-                return false;
-        }
     }
 }
